+++ CHANGELOG.md
@@ -11,8 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `#[AsTwigExtension]` attribute for declarative Twig extension registration ([63edb11], [#53])
 
+### Fixed
+
+- Fix `ViewEngineInterface` not registered in DI container for constructor injection ([1f5db42], [#57])
+
 [63edb11]: https://github.com/studiometa/foehn/commit/63edb11
 [#53]: https://github.com/studiometa/foehn/pull/53
+[1f5db42]: https://github.com/studiometa/foehn/commit/1f5db42
+[#57]: https://github.com/studiometa/foehn/pull/57
 
 ## [0.2.0] - 2026-02-05